moving the static-site-json build code to an in-repo addon
