+++ ember-cli-build.js
@@ -1,58 +1,6 @@
 'use strict';
 
-const { readFileSync } = require('fs');
-const StaticSiteJson = require('broccoli-static-site-json');
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
-const BroccoliMergeTrees = require('broccoli-merge-trees');
-const yaml = require('js-yaml');
-const { Serializer } = require('jsonapi-serializer');
-const writeFile = require('broccoli-file-creator');
-const Funnel = require('broccoli-funnel');
-const walkSync = require('walk-sync');
-const { extname } = require('path');
-
-const guidesSourcePublic = new Funnel('node_modules/@ember/guides-source/public');
-
-const VersionsSerializer = new Serializer('version', {
-  attributes: [
-    'allVersions',
-    'currentVersion',
-  ],
-});
-
-const versions = yaml.safeLoad(readFileSync('node_modules/@ember/guides-source/versions.yml', 'utf8'));
-
-let premberVersions = [];
-
-if(!process.env.JSON_ONLY) {
-  premberVersions = [...versions.allVersions, 'current'];
-}
-
-const urls = premberVersions.map(version => `/${version}`);
-
-premberVersions.forEach((premberVersion) => {
-  const filesVersion = premberVersion === 'current' ? versions.currentVersion : premberVersion;
-  const paths = walkSync(`node_modules/@ember/guides-source/guides/${filesVersion}`);
-
-  const mdFiles = paths.
-    filter(path => extname(path) === '.md')
-    .map(path => path.replace(/\.md/, ''))
-    .map(path => path.replace(/\/index$/, ''));
-
-  mdFiles.forEach((file) => {
-    urls.push(`/${premberVersion}/${file}`)
-  })
-});
-
-// setting an ID so that it's not undefined
-versions.id = 'versions';
-
-const jsonTrees = versions.allVersions.map((version) => new StaticSiteJson(`node_modules/@ember/guides-source/guides/${version}`, {
-  contentFolder: `content/${version}`,
-  type: 'contents',
-}));
-
-var versionsFile = writeFile('/content/versions.json', JSON.stringify(VersionsSerializer.serialize(versions)));
 
 module.exports = function(defaults) {
   let app = new EmberApp(defaults, {
@@ -60,9 +8,6 @@ module.exports = function(defaults) {
       'theme': 'okaidia',
       'components': ['scss', 'javascript', 'handlebars', 'http', 'json'],
       'plugins': ['line-numbers', 'normalize-whitespace']
-    },
-    prember: {
-      urls,
     },
     fingerprint: {
       extensions: ['js', 'css', 'map'],
@@ -76,5 +21,5 @@ module.exports = function(defaults) {
 
   app.import('node_modules/compare-versions/index.js');
 
-  return new BroccoliMergeTrees([app.toTree(), versionsFile, guidesSourcePublic, ...jsonTrees]);
+  return app.toTree();
 };
