+++ lib/content-guides-generator/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "content-guides-generator",
+  "keywords": [
+    "ember-addon",
+    "prember-plugin"
+  ]
+}
